cpu/esp_common/esp_riscv: improve exception handling
The multiple call of the exception handler indicates the occurrence of an exception during the handling of an exception, e.g. when the exception handler tries to access a function or data from the IROM while the cache is disabled. Previously, the exception handler was then called recursively until the stack or memory was exhausted.
This commit takes the following actions to ensure that the exception handler is not called recursively:
- The IROM is activated after waiting for the completion of possible write operations on the SPI flash.
- If there is a multiple call of the exception handler, it interrupts in the debugger if attached or stops the further execution until the WDT leads to a reset of the MCU.