.murdock: Introduce TEST_KCONFIG_DISSIMILAR_BUILD
This allows a list of applications that can differ between the kconfig
and make dependency resolution.
It first checks the make build to see if it compiles and tests if
possible, skipping if the nothing changed in the test cache.
If something fails it will escape.

Then it checks kconfig, and if the app is in the
TEST_KCONFIG_DISSIMILAR_BUILD list then it will build.
If the hash is different from make then it will run the test again,
ignoring any test cache.

This should simplify the migration, specifically when modules select
hardware implementations but fallback to software if hardware is
present, which is difficult to achieve with make.