core/macros: Add ASSERT_IS_STR_LITERAL() and SINGLEARG_OR_MULTIARG()
The function-like macro `ASSERT_IS_STR_LITERAL()` returns its argument
unmodified if it is a string literal, but fails to compile if not.

The function-like macro `SINGLEARG_OR_MULTIARG()` expands to `SINGLEARG`
if it was invoked with a single argument, or to `MULTIARG` if it was
called with multiple arguments. This is useful to dispatch a variadic
function-like macro to two different implementations. This is useful
when applying a function-like macro on the first argument of
a variadic function-like macro that may also be invoked with only one
argument. E.g. with

    #define FOO(x, ...) foo(BAR(x), __VA_ARGS__)

a call to `FOO(123)` is illegal, as there are zero variadic args given.
With `SINGLEARG_OR_MULTIARG()` this can now be implemented with full
conformance with the C standard:

    #define FOO_SINGLEARG(x) foo(BAR(x))
    #define FOO_MULTIARG(x, ...) foo(BAR(x), __VA_ARGS__)
    #define FOO(...)
        EXPAND(CONCAT(FOO_, SINGLEARG_OR_MULTIARG(__VA_ARGS__))(__VA_ARGS__))