sys/usbus: add USBUS_EVENT_USB_ACTIVE
`USBUS_EVENT_USB_ACTIVE` is generated by the control handler to inform the event handlers of interested interfaces that the USB `SET CONFIGURATION request` has been received and the EPs of the interfaces have been activated. This event can be used by the interface, for example, to prepare the first OUT data transfer.