pkg/openthread: fix error handling and synchronization
Previously, there was zero error handling in the ot_command_t commands
provided in pkg/openthread/contrib/platform_functions_wrapper.c -
ignoring the return value even of functions that are marked with
`OT_MUST_USE_RESULT`.

In addition, `ot_call_command()` never synchronized with the thread
executing the command before return the value. Given that there was
no error handling anyways, this likely was never noticed. But now that
the commands actually report an error, it is of utmost importance to
actually wait for the command to complete before checking the
commands return value.