drivers/dht: reversal of #16934 regression
This PR reintroduces the specific handling for DHT22 that was mistakenly removed by #16934 (drivers/dht: correct interpreting raw values).

The improvement in value handling for DHT11 has been preserved, except for the detection of negative values, which now occurs on the 4th byte for DHT11 instead of the 3rd byte for DHT22.

Moreover :
- improvement in reliability on ESP8266
- make it works on avr

Co-authored-by: Marian Buschsieweke <maribu@users.noreply.github.com>