sys/wait: add module for backend agnostic waits
This adds to sets of delay functions. The first kind consists of

- `wait_us()`: delay execution in units of microseconds
- `wait_ms()`: delay execution in units of milliseconds
- `wait_s()`: delay execution in units of seconds

These will use `ztimer` as backend, if used anyways. Otherwise a
trivial CPU counting loop is used.

The second kind consists of

- `wait_us_spinning()`: busy-wait for a given number of microseconds

The benefit of this compared to `ztimer_spin()` is that platform
specific implementation can be used to increase accuracy. This can
allow timing critical drivers (e.g. bit-banging with accuracy
requirements with in the order of a few microseconds) to scale down
to slow MCUs.