Makefile.include: turn CFLAGS into a sane variable
Variable in Makefiles are by default insane (a.k.a. recursively
expended variables). However, sane variables (a.k.a. simply expended
variables) are clearly better.

This assigns `CFLAGS` the empty string using a simple assignment.
Subsequent use of `+=` to extend `CFLAGS` will keep it a sane. This
avoids calling shell functions generating the `CPU_RAM_SIZE` macro
for each and every module that is compiled.