cpu/avr8_common: Optimize context ISR switch
The current context switch implementaion don't take in consideration
the call-used registers.  This means that ISR and context switch can
not be optimized.  This rework both context switch and ISR save/restore
order to optimize the context switch on ISR and reduce code size.

This chance should improve the performance of AVR platform on RIOT-OS
if is compared with the 2023.04 release.

Signed-off-by: Gerson Fernando Budke <nandojve@gmail.com>