pkg/arduino_adafruit_sensor: fix dependencies
This fixes the dependencies of the `arduino_adafruit_sensor` package,
which previously relied on the `arduino` feature. This feature no longer
exists, as it was split into more fine granular features. However, the
module should never have used that feature directly in the first place,
but rather just use the arduino module. This in turn depends on the
correct features.

(cherry picked from commit acc1203606302fbccc82cfa372c400a342d28976)