nanocoap: Make coap_block_finish more resilient
This API change removes the error prone nature of coap_block_finish and
related functions. Before this change the position of the block-wise
options in the coap packet buffer was not allowed to change or the
coap_block_finish function would update the wrong byte in the packet
buffer.

After this change the block option is looked up dynamically from the
packet buffer, ensuring that if the position got changed the function
would still behave correctly.